web-view/requirements.txt
flask==2.1.2
psycopg2-binary==2.9.1
